tests/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    update_tools
#
absl-py==0.8.1            # via tensorboard, tensorflow
astor==0.8.0              # via tensorflow
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
azure-common==1.1.23
azure-core==1.1.1         # via azure-storage-blob
azure-nspkg==3.0.2        # via azure-common, azure-core, azure-storage-nspkg
azure-storage-blob==12.1.0
azure-storage-nspkg==3.1.0  # via azure-storage-blob
backports.weakref==1.0.post1  # via tensorflow
cachetools==3.1.1         # via google-auth
certifi==2019.6.16        # via msrest, requests
cffi==1.13.2              # via cryptography
chardet==3.0.4            # via requests
configparser==4.0.2       # via importlib-metadata
contextlib2==0.5.5        # via importlib-metadata
cryptography==2.8         # via azure-storage-blob
dill==0.2.9               # via google-gax
enum34==1.1.6             # via absl-py, azure-core, azure-storage-blob, cryptography, google-cloud-language, grpcio, msrest, tensorflow
funcsigs==1.0.2           # via mock, pytest
functools32==3.2.3.post2  # via tensorflow
future==0.16.0            # via google-gax
futures==3.3.0            # via azure-storage-blob, google-cloud-core, grpcio, tensorboard
gast==0.2.2               # via tensorflow
google-auth==1.6.3        # via google-cloud-core, google-gax
google-cloud-core==0.27.1  # via google-cloud-language
google-cloud-language==0.29.0
google-gax==0.15.16       # via google-cloud-language
google-pasta==0.1.7       # via tensorflow
googleapis-common-protos[grpc]==1.6.0  # via google-cloud-core, google-cloud-language, google-gax
grpcio==1.23.0            # via google-gax, googleapis-common-protos, tensorboard, tensorflow
h5py==2.10.0              # via keras-applications
idna==2.8                 # via requests
importlib-metadata==0.22  # via pluggy
ipaddress==1.0.23         # via cryptography
isodate==0.6.0            # via msrest
keras-applications==1.0.8  # via tensorflow
keras-preprocessing==1.1.0  # via tensorflow
markdown==3.1.1           # via tensorboard
mock==3.0.5
more-itertools==5.0.0     # via pytest, zipp
msrest==0.6.10            # via azure-storage-blob
numpy==1.16.5             # via h5py, keras-applications, keras-preprocessing, opt-einsum, tensorboard, tensorflow
oauthlib==3.1.0           # via requests-oauthlib
opt-einsum==2.3.2         # via tensorflow
pathlib2==2.3.4           # via importlib-metadata, pytest
pg8000==1.12.5            # via testing.postgresql
pluggy==0.13.0            # via pytest
ply==3.8                  # via google-gax
protobuf==3.9.1           # via google-cloud-core, google-gax, googleapis-common-protos, tensorboard, tensorflow
py==1.8.0                 # via pytest
pyasn1-modules==0.2.6     # via google-auth
pyasn1==0.4.7             # via pyasn1-modules, rsa
pycparser==2.19           # via cffi
pymysql==0.9.3            # via testing.mysqld
pytest==3.10.1
requests-oauthlib==1.3.0  # via msrest
requests==2.22.0          # via azure-core, google-cloud-core, google-gax, msrest, requests-oauthlib
rsa==4.0                  # via google-auth
scandir==1.10.0           # via pathlib2
six==1.12.0               # via absl-py, azure-core, cryptography, google-auth, google-cloud-core, grpcio, h5py, isodate, keras-preprocessing, mock, more-itertools, pathlib2, pg8000, protobuf, pytest, tensorboard, tensorflow
tensorboard==2.0.0        # via tensorflow
tensorflow-estimator==2.0.1  # via tensorflow
tensorflow==2.0.0
termcolor==1.1.0          # via tensorflow
testing.common.database==2.0.3  # via testing.mysqld, testing.postgresql
testing.mysqld==1.4.0
testing.postgresql==1.3.0
typing==3.7.4.1           # via azure-core, azure-storage-blob, msrest
urllib3==1.25.3           # via requests
werkzeug==0.16.0          # via tensorboard
wheel==0.33.6             # via tensorboard, tensorflow
wrapt==1.11.2             # via tensorflow
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
setuptools==41.0.1        # via google-cloud-core, markdown, protobuf, pytest, tensorboard
